feat: merge create origin post & vote post file (#20)
